split event filtering & testing logic into new files
